Move `np` to devDependencies
The np dependency is only used in the release scripts and should not be
marked as an optional runtime dependency.

Having np be installed as an optional runtime dependency is causing
warnings on npm and yarn installation of the eslint-formatter-summary
package due to issues in the dependencies of np such as security
vulnerabilities and warnings about the `engines` block defined by np.